requirements.txt
django==2.2.8
djangorestframework
django-heroku
djongo
dnspython
pillow